edit form handles own submission
